move all activities to single package
