+++ app/src/main/java/org/ahomewithin/ahomewithin/activities/ServicesActivity.java
@@ -1,4 +1,4 @@
-package org.ahomewithin.ahomewithin.services;
+package org.ahomewithin.ahomewithin.activities;
 
 import android.content.Context;
 import android.os.Bundle;
@@ -7,11 +7,9 @@ import android.view.View;
 import android.widget.FrameLayout;
 
 import org.ahomewithin.ahomewithin.AHomeWithinClient;
-import org.ahomewithin.ahomewithin.MainActivity;
 import org.ahomewithin.ahomewithin.R;
 import org.ahomewithin.ahomewithin.models.Event;
 import org.ahomewithin.ahomewithin.models.Provider;
-import org.ahomewithin.ahomewithin.models.Recommended;
 import org.json.JSONObject;
 
 import java.util.List;
